ray, canvas ppm output

+++ Util.cs
@@ -7,7 +7,7 @@ namespace ray_tracer
     public class Util
     {
         public const float EPSILON = 0.00001f;
-        public static bool NearlyEqual(float a, float b) => Math.Abs(a - b) < EPSILON;
+        public static bool NearlyEqual(double a, double b) => Math.Abs(a - b) < EPSILON;
 
         private Util() { }
     }
